+++ packages/orq-rc/src/models/operations/updatebudget.ts
@@ -547,7 +547,7 @@ export const UpdateBudgetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-10-28T12:57:47.340Z",
+    "2025-10-28T15:44:09.367Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -585,7 +585,7 @@ export const UpdateBudgetResponseBody$outboundSchema: z.ZodType<
   isActive: z.boolean(),
   consumption: z.lazy(() => UpdateBudgetConsumption$outboundSchema).optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-10-28T12:57:47.340Z"))
+  updated: z.date().default(() => new Date("2025-10-28T15:44:09.367Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {